feat: Implemented caluculation of sum of squares
The sumSquares function takes two vectors as input and calucluates the sum of the squares of the two vectors.
The program will also print out the average values for x and y as well as the sum of the squares.
Refactored some of the print statements to use string literals for easier reading.