feat: add compatability with csv files
The script will now allow you to select the csv file you wish to use.
It will select the first two columns for x and y for now.
If there is a header row, then the script will ignore it.
TODO: create a GUI and reimplement user input to make it compatible.