Update the functionality of the buttons in the initial window.
Now if you press "Next" without an input, it will open a popup that will inform the user to select a file.
If you press "Next" with a file selected, it will close the window.
If you hit the manual input button, it will now close the window.
Closing the window will still act the same way.